Add more features and refactor code in script.js.
User can now increase the opacity of an element of
the grid when mouseenter event is fired on said
element. Each time the event is fired the
background color of an element is randomized.
Said event is only executing when LMB is actively
being pressed.
Improve style on the index page.